Factor helpers out and reuse for chgrp.
